OOPS, I was wrong. the issue wasnt in setting the normal vector to the difference between closest point and AtoB vector, RATHER, you can use either of those vectors, because their direction is the same.
The issue was that, whichever of those two you use, you must NORMALIZE the final value before setting it as the body normal.

Now the only issue is that of units. the extents of the cube are not being calculated correctly, so the cube is being penetrated. need to change the cube side calculations.

line 214 and 306